`defmt-rtt`: Move `struct Channel` into own module
